+++ Configs/Linux/Debian/fjrodafo/.config/eww/eww.bash
@@ -4,8 +4,8 @@
 #  ║╣ │  ╠╩╗│ ││││├─┤├┬┘└─┐  ║║║├─┤│  ├┴┐└┬┘  ║║║│ │││ ┬├┤  │ └─┐
 #  ╚═╝┴─┘╩ ╩└─┘└┴┘┴ ┴┴└─└─┘  ╚╩╝┴ ┴└─┘┴ ┴ ┴   ╚╩╝┴─┴┘└─┘└─┘ ┴ └─┘
 
-# This file launch EWW widgets
-launch_eww_widgets(){
+# This file launch EWW
+launch_eww(){
     # Check if EWW bar/calendar is active by using 'eww active-windows'
     local eww_active_windows=$(~/.eww/target/release/eww active-windows)
 
@@ -25,4 +25,4 @@ launch_eww_widgets(){
     bspc config left_padding $( [[ "$eww_active_windows" == *"bar"* ]] && echo 0 || echo 54 )
 }
 
-launch_eww_widgets
+launch_eww