+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ```
+* Commit: 116ea33 - v156
+| Author: fjrodafo - Fri, 24 Jan 2025 21:40:00 +0000
+| 
 * Commit: 27c806d - v155
 | Author: fjrodafo - Fri, 24 Jan 2025 10:56:47 +0000
 |